+++ src/includes/admin/helpers/class-alma-order-helper.php
@@ -28,13 +28,12 @@ class Alma_Order_Helper {
 	 *
 	 * @return \WC_Order[]
 	 */
-	public static function get_orders_by_date_range( $from, $to ) {
-		$args = array(
-			'date_created' => $from . '...' . $to,
-			'type'         => 'shop_order',
+	public function get_orders_by_date_range( $from, $to ) {
+		return wc_get_orders(
+			array(
+				'date_created' => $from . '...' . $to,
+				'type'         => 'shop_order',
+			)
 		);
-
-		return wc_get_orders( $args );
 	}
-
 }