+++ src/includes/helpers/class-alma-checkout-helper.php
@@ -71,7 +71,7 @@ class Alma_Checkout_Helper {
 
 		$this->logger->error(
 			sprintf(
-				'Nonce not found or wrong - FieldName "%s" - NonceName "%s" - FieldName value "%s" - NonceName  value"%s"',
+				'Nonce not found or wrong - FieldName "%s" - NonceName "%s"  - data "%s"',
 				$field_name,
 				$nonce_name,
 				wp_json_encode( $_POST )